Renames CachedObjectStorage to Collection/Cells
