+++ src/PhpSpreadsheet/Collection/Cells/MemoryGZip.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace PhpOffice\PhpSpreadsheet\CachedObjectStorage;
+namespace PhpOffice\PhpSpreadsheet\Collection\Cells;
 
 /**
  * Copyright (c) 2006 - 2016 PhpSpreadsheet.
@@ -24,7 +24,7 @@ namespace PhpOffice\PhpSpreadsheet\CachedObjectStorage;
  * @copyright  Copyright (c) 2006 - 2016 PhpSpreadsheet (https://github.com/PHPOffice/PhpSpreadsheet)
  * @license    http://www.gnu.org/licenses/old-licenses/lgpl-2.1.txt    LGPL
  */
-class Igbinary extends CacheBase implements ICache
+class MemoryGZip extends CacheBase implements ICache
 {
     /**
      * Store cell data in cache for the current cell object if it's "dirty",
@@ -37,13 +37,11 @@ class Igbinary extends CacheBase implements ICache
         if ($this->currentCellIsDirty && !empty($this->currentObjectID)) {
             $this->currentObject->detach();
 
-            $this->cellCache[$this->currentObjectID] = igbinary_serialize($this->currentObject);
+            $this->cellCache[$this->currentObjectID] = gzdeflate(serialize($this->currentObject), 9);
             $this->currentCellIsDirty = false;
         }
         $this->currentObjectID = $this->currentObject = null;
     }
-
-    //    function _storeData()
 
     /**
      * Add or Update a cell in cache identified by coordinate address.
@@ -92,15 +90,13 @@ class Igbinary extends CacheBase implements ICache
 
         //    Set current entry to the requested entry
         $this->currentObjectID = $pCoord;
-        $this->currentObject = igbinary_unserialize($this->cellCache[$pCoord]);
+        $this->currentObject = unserialize(gzinflate($this->cellCache[$pCoord]));
         //    Re-attach this as the cell's parent
         $this->currentObject->attach($this);
 
         //    Return requested entry
         return $this->currentObject;
     }
-
-    //    function getCacheData()
 
     /**
      * Get a list of all cell addresses currently held in cache.
@@ -129,22 +125,5 @@ class Igbinary extends CacheBase implements ICache
 
         //    detach ourself from the worksheet, so that it can then delete this object successfully
         $this->parent = null;
-    }
-
-    //    function unsetWorksheetCells()
-
-    /**
-     * Identify whether the caching method is currently available
-     * Some methods are dependent on the availability of certain extensions being enabled in the PHP build.
-     *
-     * @return bool
-     */
-    public static function cacheMethodIsAvailable()
-    {
-        if (!function_exists('igbinary_serialize')) {
-            return false;
-        }
-
-        return true;
     }
 }
